Lay out a foundation for integration tests
This adds pytest-docker as a pytest plugin that allows to add a docker service as a fixture for test cases.

Parameterizes the Superset version, and allows to not load example data which takes a very long time.

Adds a very simeple test case, just to confirm that the integration test setup works.